+++ src/frontend/web/src/components/Chat/Input.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import styled from 'styled-components';
-import sendIcon from '../icon/send.svg';
-import BoldIcon from '../icon/bold.svg';
-import ItalicIcon from '../icon/italic.svg';
-import StrikeIcon from '../icon/strike.svg';
-import CodeIcon from '../icon/code.svg';
-import ListBulletIcon from '../icon/list_bullet.svg';
-import ListNumberIcon from '../icon/list_number.svg';
-import LinkIcon from '../icon/link.svg';
-import CodeBlockIcon from '../icon/code_block.svg';
-import QuoteIcon from '../icon/quote.svg';
-import EmailIcon from '../icon/email.svg';
-import SmileIcon from '../icon/smile.svg';
-import FileIcon from '../icon/file.svg';
+import sendIcon from '../../icon/send.svg';
+import BoldIcon from '../../icon/bold.svg';
+import ItalicIcon from '../../icon/italic.svg';
+import StrikeIcon from '../../icon/strike.svg';
+import CodeIcon from '../../icon/code.svg';
+import ListBulletIcon from '../../icon/list_bullet.svg';
+import ListNumberIcon from '../../icon/list_number.svg';
+import LinkIcon from '../../icon/link.svg';
+import CodeBlockIcon from '../../icon/code_block.svg';
+import QuoteIcon from '../../icon/quote.svg';
+import EmailIcon from '../../icon/email.svg';
+import SmileIcon from '../../icon/smile.svg';
+import FileIcon from '../../icon/file.svg';
 
 const InputContainer = styled.article`
   flex-shrink: 0;
